:recycle: moved the coin back to data module, changed the map syntax
