+++ helpers/pyband/pyband/data.py
@@ -5,22 +5,31 @@ from typing import List, Optional, NewType
 from dacite import Config
 from pyband.utils import parse_epoch_time
 from pyband.wallet import Address
-from pyband.coin import Coin
 
 HexBytes = NewType("HexBytes", bytes)
 EpochTime = NewType("EpochTime", int)
 
 
-DACITE_CONFIG = Config(
-    type_hooks={
-        int: int,
-        bytes: base64.b64decode,
-        HexBytes: bytes.fromhex,
-        EpochTime: parse_epoch_time,
-        Address: Address.from_acc_bech32,
-        Coin: Coin.from_json,
-    }
-)
+@dataclass
+class Coin(object):
+    amount: int
+    denom: str
+
+    @classmethod
+    def from_json(cls, coin) -> "Coin":
+        return cls(int(coin["amount"]), coin["denom"])
+
+    def as_json(self) -> dict:
+        return {"amount": str(self.amount), "denom": self.denom}
+
+    def validate(self) -> bool:
+        if self.amount < 0:
+            raise ValueError("Expect amount more than 0")
+
+        if len(self.denom) == 0:
+            raise ValueError("Expect denom")
+
+        return True
 
 
 @dataclass
@@ -180,3 +189,15 @@ class ReferencePrice(object):
     pair: str
     rate: float
     updated_at: ReferencePriceUpdated
+
+
+DACITE_CONFIG = Config(
+    type_hooks={
+        int: int,
+        bytes: base64.b64decode,
+        HexBytes: bytes.fromhex,
+        EpochTime: parse_epoch_time,
+        Address: Address.from_acc_bech32,
+        Coin: Coin.from_json,
+    }
+)
